+++ src/network/send.h
@@ -639,6 +639,7 @@ public:
  *
  *	Packet 0x70 : PacketEffect				displays a visual effect (NORMAL)
  *	Packet 0xC0 : PacketEffect				displays a hued visual effect (NORMAL)
+ *	Packet 0xC7 : PacketEffect				displays a particle visual effect (NORMAL)
  *
  *
  ***************************************************************************/
@@ -647,7 +648,7 @@ class PacketEffect : public PacketSend
 public:
 	PacketEffect(const CClient* target, EFFECT_TYPE motion, ITEMID_TYPE id, const CObjBaseTemplate* dst, const CObjBaseTemplate* src, BYTE speed, BYTE loop, bool explode);
 	PacketEffect(const CClient* target, EFFECT_TYPE motion, ITEMID_TYPE id, const CObjBaseTemplate* dst, const CObjBaseTemplate* src, BYTE speed, BYTE loop, bool explode, DWORD hue, DWORD render);
-	PacketEffect(const CClient* target, EFFECT_TYPE motion, ITEMID_TYPE id, const CObjBaseTemplate* dst, const CObjBaseTemplate* src, BYTE speed, BYTE loop, bool explode, DWORD hue, DWORD render, WORD effectid, DWORD explodeid, WORD explodesound, DWORD effectuid, BYTE type);
+	PacketEffect(const CClient* target, EFFECT_TYPE motion, ITEMID_TYPE id, const CObjBaseTemplate* dst, const CObjBaseTemplate* src, BYTE speed, BYTE loop, bool explode, DWORD hue, DWORD render, WORD effectid, WORD explodeid, WORD explodesound, DWORD effectuid, BYTE type);
 	void writeBasicEffect(EFFECT_TYPE motion, ITEMID_TYPE id, const CObjBaseTemplate* dst, const CObjBaseTemplate* src, BYTE speed, BYTE loop, bool explode);
 	void writeHuedEffect(DWORD hue, DWORD render);
 };
